Habitling Card component updated
