+++ my-react/src/App.jsx
@@ -1,15 +1,13 @@
 import './App.css';
 // import Check_habit from './Check_habit'
-import NavBar from './Components/NavBar'
-import Habitling from './Components/Habitling'
-import HabitCreate from './Components/HabitCreate'
+import NavBar from './Components/NavBar';
+import Habitling from './Components/Habitling';
 function App() {
 
   return (
     <>
       <NavBar/>
       <Habitling/>
-      <HabitCreate/>
     </>
   );
 };
