+++ my-react/src/Components/Habitling.css
@@ -1,4 +1,4 @@
-.box {
+.habitling-box {
     border: 2px solid black;
     padding: 0px;
     border-radius: 10px;
@@ -9,7 +9,7 @@
     text-align: center;
 }
 
-.name {
+.habitling-name {
     align-items: center;
     justify-content: center;
     border-bottom: 2px solid black;
@@ -20,9 +20,16 @@
     margin: 0 auto;  
 }
 
+.habitling-description {
+    color: #666;
+    margin-top: 8px;
+    align-items: center;
+    justify-content: center;
+    border-bottom: 2px solid black;
+    margin: 0 auto;
+}
 
-
-.stats-box {
+.habitling-stats-box {
     display: flex;
     margin-top: 10px;
     align-items: center;
@@ -31,64 +38,84 @@
     flex-direction: column;
 }
 
-.streak {
+.habitling-stats-streak {
     color: green;
 }
 
-.streak-best {
+.habitling-stats-best {
     color: blue;
 }
 
-.pet-box {
-    height: 200px;
-    position: relative;
-}
-
-.sprite{
-    position: absolute;
-    top:37%;
-    left: 35%;
+.habitling-pet-box img {
+    width: 60px;
+    height: 60px;
+    border-radius: 50%;
+    margin-top: 10px;
 }
 
 
 
 
+/*--------------------------------*/
 
-/* check box's olds css */
-form{
+
+.habit-container {
     display: flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    justify-content: space-between;
-    border-bottom: 2px solid black;
-    /* border: 2px solid black; */
-    width:100%;
-
-
-
-}
-.checkbox-description{
-    width:80%;
-    padding-left: 2rem;
-    text-align:left;
-        
-}
-.habit-interval{
-
-    padding-left: 2rem;
-}
-.check-habit{
-    display: flex;
-    flex-direction: column ;
     justify-content: center;
-    align-content: center;
-    padding:4%;
-}
-.check-habit input{
-    width: 3rem; 
-    height: 3rem;
-    background-color: #007bff;
-    color: white;
-    border-radius: 8px;
-    cursor: pointer;
-}
+    gap: 20px;
+    padding: 20px;
+    flex-wrap: wrap;
+  }
+  
+  .habit-card {
+    color: black;
+    width: 200px;
+    background: white;
+    border-radius: 15px;
+    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.2);
+    padding: 15px;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .habit-title {
+    font-size: 18px;
+    font-weight: bold;
+  }
+  
+  .habit-label {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    margin-top: 10px;
+  }
+  
+  .habit-checkbox {
+    width: 16px;
+    height: 16px;
+  }
+  
+  .habit-frequency {
+    font-size: 14px;
+    margin-top: 5px;
+  }
+  
+  .habit-streak {
+    margin-top: 10px;
+    font-size: 14px;
+  }
+  
+  .habit-image-container {
+    margin-top: 15px;
+  }
+  
+  .habit-image {
+    width: 60px;
+    height: 60px;
+    border-radius: 10px;
+    background: lightblue;
+    padding: 10px;
+  }
+
