+++ my-react/src/Habitling.css
@@ -1,4 +1,4 @@
-/* .habitling-box {
+.habitling-box {
     border: 2px solid black;
     padding: 0px;
     border-radius: 10px;
@@ -51,70 +51,4 @@
     height: 60px;
     border-radius: 50%;
     margin-top: 10px;
-} */
-
-
-
-
-/*--------------------------------*/
-
-
-.habit-container {
-    display: flex;
-    justify-content: center;
-    gap: 20px;
-    padding: 20px;
-    flex-wrap: wrap;
-  }
-  
-  .habit-card {
-    width: 200px;
-    background: white;
-    border-radius: 15px;
-    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.2);
-    padding: 15px;
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .habit-title {
-    font-size: 18px;
-    font-weight: bold;
-  }
-  
-  .habit-label {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    margin-top: 10px;
-  }
-  
-  .habit-checkbox {
-    width: 16px;
-    height: 16px;
-  }
-  
-  .habit-frequency {
-    font-size: 14px;
-    margin-top: 5px;
-  }
-  
-  .habit-streak {
-    margin-top: 10px;
-    font-size: 14px;
-  }
-  
-  .habit-image-container {
-    margin-top: 15px;
-  }
-  
-  .habit-image {
-    width: 60px;
-    height: 60px;
-    border-radius: 10px;
-    background: lightblue;
-    padding: 10px;
-  }
-  
+}
